+++ NEWS.md
@@ -9,6 +9,7 @@
 * Added optional low-pass filtering of motion parameters, matching (Gratton)
 * Gracefully adjust motion filtering parameters if they fall above Nyquist at this TR
 * Modify extract ROIs config to avoid input_regex and always generate it internally from postproc stream
+* Add optional header row for postprocessed confounds TSVs, configurable via postprocess YAML and validated during setup
 
 # BrainGnomes 0.7-5